chore: 폴더구조 정리
