+++ SaveEarth/SaveEarth/Manager/Weather/Data/AppLocation.swift
