+++ SaveEarth/SaveEarth/Manager/Weather/Model/TemperatureData.swift
