+++ SaveEarth/SaveEarth/Manager/Weather/WeatherManager.swift
